AdditionLevelWorking
not completely, but im just about done with one level being able to work. I will copy and paste the code to the rest of the levels when i get the chance and give the buttons some functionality when you click the answer choices. Then we are with that. Just need to add some counters for the stats and maybe change up the ui and shit.